backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-listing-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0,<0.113.0",
    "uvicorn[standard]>=0.30.0,<0.32.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "alembic>=1.13.0,<2.0.0",
    "pydantic[email]>=2.7.0,<3.0.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    "structlog>=24.1.0,<25.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "python-multipart>=0.0.9,<0.1.0",
]

[project.optional-dependencies]
# Install for PostgreSQL in production: pip install -e ".[postgres]"
postgres = [
    "psycopg2-binary>=2.9.9,<3.0.0",
]
dev = [
    "pytest>=8.2.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "httpx>=0.27.0,<0.28.0",
    "ruff>=0.4.0,<0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
